requirements.txt
django
sqlalchemy 
asyncpg 
uvicorn
python-dotenv
psycopg[binary]
djangorestframework